dscnitrogen
the chances that you will be able to actually search the nitro are extremely small. In addition, you will still get a cooldown from the discord when sending a webhook request. but just for fun, why not. if you have any problems because of this, I'm not responsible for them.